Allow removing and reinserting pkcs11 keys.
This introduces a check in pkcs11-helper verifying the
smart card is ready to sign data (was not removed since last invocation). If
it was, the new code tries to ask a user for the PIN (using askpass program
if defined) and reauthenticate the card, rather than fail hard.

If you're using this with OpenSC, you need version 0.20.0 or above.

Originally by Jakub Jelen <jjelen@redhat.com> and
Nikos Mavrogiannopoulos <nmav@redhat.com>. Brought up to date as of
2020-03-17 by Jacob Hoffman-Andrews <jsha@letsencrypt.org>.